Worked on Synthesizer programming and readback

+++ rx63n/bert100/xo.h
@@ -1,1 +1,1 @@
-void XO_Init(const char *name,uint8_t i2cAddr);
+void XO_Init(const char *name,uint16_t i2cAddr);
